+++ app/native-components/Header/constants.ts
@@ -1,0 +1,137 @@
+export const NAVLINKS = [
+    {
+        href: "/",
+        label: "Home",
+    },
+    {
+        href: "/about",
+        label: "About Us",
+    },
+    {
+        href: "/catalogue",
+        label: "Download Catalogue",
+    },
+    {
+        href: "/contact",
+        label: "Contact Us",
+    },
+];
+
+export const PRODUCT_CATEGORIES = [
+    {
+        label: "Ceiling Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "Chandeliers",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+            {
+                label: "Pendant Lights",
+                href: "",
+                imageHref: "/global/random-image-2.jpg",
+            },
+            {
+                label: "Recessed Lighting",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+        ],
+    },
+    {
+        label: "Wall Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "Wall Sconces",
+                href: "",
+                imageHref: "/global/random-image-2.jpg",
+            },
+            {
+                label: "Picture Lights",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+            {
+                label: "Bedside Wall Lamps",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+        ],
+    },
+    {
+        label: "Outdoor Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "Garden Path Lights",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+            {
+                label: "Wall-Mounted Outdoor Lights",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+        ],
+    },
+    {
+        label: "Smart Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "Smart Bulbs",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+            {
+                label: "Voice-Controlled Lighting Systems",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+            {
+                label: "Motion Sensor Lights",
+                href: "",
+                imageHref: "/global/random-image-2.jpg",
+            },
+        ],
+    },
+    {
+        label: "Architectural Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "Linear LED Systems",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+            {
+                label: "Track Lighting",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+        ],
+    },
+    {
+        label: "Commercial & Industrial Lighting",
+        href: "",
+        subcategories: [
+            {
+                label: "High Bay Lights",
+                href: "",
+                imageHref: "/global/random-image-3.jpg",
+            },
+            {
+                label: "Emergency Exit Lights",
+                href: "",
+                imageHref: "/global/random-image-1.jpg",
+            },
+            {
+                label: "Office Panel Lights",
+                href: "",
+                imageHref: "/global/random-image-2.jpg",
+            },
+        ],
+    },
+];
